[webui] Remove list_simple action
The list simple action was only used for the case of browsers having
no javascript support and spiders. The latter case is already covered
by a condition in the list action.
The javascript case is not supported. So we can simply remove the action
and routing and only keep the list_simple view