[api][webui] Rename DownloadRepository to DODSource
Download reposiories are actually just a representation of a source of package
binaries that can be added to a repository. A (DoD) Repository can have more than
one DoD source.

This changes aim for making the difference between DoD sources and repositories
more obvious. But is also a preparation for adding a DoD repository class that
inherits from Repository model.