[ci] Rails 5, capybara: Explicitly search for invisble element
The element we select in our test, and we later on check for a changed label,
is invisible until the user hovers over it. To do so in our tests we have to
find() and then hover() the element.

With the recent capybara update we need to explicitly set the visible: false
parameter when calling find().