[frontend] Add js style rule: Require triple equals
Use === instead of ==. The == and != operator are not only comparing two
values. But also trying to convert the variable on the right side, if their
types are not identical. This can be quite a complex logic due to the
various combinations[*] that are possible.

To simplify things and avoid confusion for people that are js experts,
who is?, it's better to stick with ===.

https://dorey.github.io/JavaScript-Equality-Table/