[frontend] Make Backend::Api::Cloud#upload_jobs independent from user
The method was returning backend data of a all upload_jobs of a given user.
Instead it now receives a list of job ids and returns their associated
backend data.

This makes the method more flexible and allows it to be used in
different contexts, eg. querying data of a single upload job.