[ci] Fix file upload in travis failure script
When our test suite fails we upload our frontend log files to
https://transfer.sh. For that we are using a relativ path.
Since during the test preparation we might switch directories and
we can not predict when our test suite fail we have to use a static
path.
This avoids that we upload other files than intended due to failed
attempts of switching to the log dir.