[ci] Remove git-cop
We've been using git-cop to ensure a certain format of our commit
messages. Part of that was to ensure that one of the OBS tags
were used and the length of the commit message.
Most of the times commits to OBS are applying to this format anyway
and both of this can also be ensured via pull request reviews.
Given the amount of effort we have to take to keep git-cop up to
date it's better to check the commit message format via PR reviews.