[dist] Rpmlint: Silence devel-file-in-non-devel-package warnings
Since we soon switch to the bundle gems service we need to solve
some rpmlint issues caused by the new gem files we store in the
vendor/ directory.

Many of the bundled gems we package contain C header files.
Since we won't be able to solve all these cases upstream in
reasonable time we disable that check for now.