Add groups to user image tag helper and use keyword arguments
We will show the reviewers gravatar on a request in the
staging workflow. Since groups can be reviewers as well, we
need to be able to show the gravatar of a group.
In order to show the name of the reviewer i added
a title to the user and group image tag helper. Since the helper
takes quite a lot of arguments i used keyword arguments.