Rewrite querying incident packages
Replaces the old hash mapping code with a SQL query.

The previous code was also checking whether the package name had a
suffix that matched the release target, e.g. my_package.openSUSE_42.
This was dropped, since we already ensure that the package has a build
flag explicitly enabled for one of the release targets.